Refactor memory and SQL backends
